requirements.txt
pyzbar
numpy
openai
pyttsx3
pyaudio
pyserial
Pillow